[cw] Save woker output in s3 bucket
To be able to do a distributed work, the worker will save it's output
file (after final parser) in the s3 bucket.
After a distributed work, the master will gather all files in the s3
folder that has the output files of the workers and put it into a single
file.

Probably, in the future, there will be a new worker that will be
schedule to check if there's a new file available to append it to the
total result